kubeflow/requirements.txt
# Kubeflow Pipelines SDK
kfp>=2.7.0
kfp-pipeline-spec>=0.2.0

# Kubernetes client for KServe
kubernetes>=28.0.0

# Kafka integration
confluent-kafka>=2.3.0

# Model handling
transformers>=4.35.0
torch>=2.0.0
huggingface-hub>=0.19.0

# Utilities
pydantic>=2.5.0
pyyaml>=6.0.1
requests>=2.31.0
